Connected Modal to Backend, Removed ask Question page
